requirements.txt
beautifulsoup4==4.5.3
requests==2.13.0
python-twitter
